+++ R/get_elev_raster.R
@@ -13,10 +13,12 @@
 #'           details on zoom and resolution see the documentation from Mapzen at 
 #'           \url{https://github.com/tilezen/joerd/blob/master/docs/data-sources.md#what-is-the-ground-resolution}.
 #'           The z is not required for the OpenTopography data sources. 
-#' @param prj A PROJ.4 string defining the projection of the locations argument. 
-#'            If a \code{sp} or \code{raster} object is provided, the PROJ.4 
-#'            string will be taken from that.  This argument is required for a 
-#'            \code{data.frame} of locations."
+#' @param prj A string defining the projection of the locations argument. The 
+#'            string needs to be an acceptable SRS_string for 
+#'            \code{\link[sp]{CRS}} for your version of PROJ. If a \code{sf} 
+#'            object, a \code{sp} object or a \code{raster} object 
+#'            is provided, the string will be taken from that.  This 
+#'            argument is required for a \code{data.frame} of locations.
 #' @param src A character indicating which API to use.  Currently supports "aws" 
 #'            and "gl3", "gl1", "alos", or "srtm15plus" from the OpenTopography API global 
 #'            datasets. "aws" is the default.
@@ -64,6 +66,8 @@
 #' @importFrom sp wkt
 #' @examples 
 #' \dontrun{
+#' data(lake)
+#' 
 #' loc_df <- data.frame(x = runif(6,min=sp::bbox(lake)[1,1], 
 #'                                max=sp::bbox(lake)[1,2]),
 #'                      y = runif(6,min=sp::bbox(lake)[2,1], 
@@ -71,7 +75,9 @@
 #' # Example for PROJ > 5.2.0
 #' x <- get_elev_raster(locations = loc_df, prj = sp::wkt(lake) , z=10)
 #' 
-#' data(lake)
+#' # Example for PROJ < 5.2.0 
+#' x <- get_elev_raster(locations = loc_df, prj = sp::proj4string(lake) , z=10)
+
 #' x <- get_elev_raster(lake, z = 12)
 #' x <- get_elev_raster(lake, src = "gl3", expand = 5000)
 #' }
@@ -126,7 +132,7 @@ get_elev_raster <- function(locations, z, prj = NULL,
   }
  
   if(verbose){
-    message(paste("Note: Elevation units are in meters.\nNote: The coordinate reference system is:\n", prj))
+    message(paste("Note: Elevation units are in meters."))
   }
   
   
@@ -154,7 +160,12 @@ get_elev_raster <- function(locations, z, prj = NULL,
 #'          of the resultant raster is determined by the zoom and latitude.  For 
 #'          details on zoom and resolution see the documentation from Mapzen at 
 #'          \url{https://github.com/tilezen/joerd/blob/master/docs/data-sources.md#what-is-the-ground-resolution}
-#' @param prj PROJ.4 string for input bbox 
+#' @param prj A string defining the projection of the locations argument. The 
+#'            string needs to be an acceptable SRS_string for 
+#'            \code{\link[sp]{CRS}} for your version of PROJ. If a \code{sf} 
+#'            object, a \code{sp} object or a \code{raster} object 
+#'            is provided, the string will be taken from that.  This 
+#'            argument is required for a \code{data.frame} of locations. 
 #' @param expand A numeric value of a distance, in map units, used to expand the
 #'               bounding box that is used to fetch the terrain tiles. This can 
 #'               be used for features that fall close to the edge of a tile and 
@@ -291,7 +302,12 @@ merge_rasters <- function(raster_list,  target_prj, method = "bilinear", returnR
 #' 
 #' @param locations Either a \code{data.frame} of x (long) and y (lat), an 
 #'                  \code{sp}, an \code{sf}, or \code{raster} object as input. 
-#' @param prj PROJ.4 string for input bbox 
+#' @param prj A string defining the projection of the locations argument. The 
+#'            string needs to be an acceptable SRS_string for 
+#'            \code{\link[sp]{CRS}} for your version of PROJ. If a \code{sf} 
+#'            object, a \code{sp} object or a \code{raster} object 
+#'            is provided, the string will be taken from that.  This 
+#'            argument is required for a \code{data.frame} of locations. 
 #' @param expand A numeric value of a distance, in map units, used to expand the
 #'               bounding box that is used to fetch the SRTM data. 
 #' @param ... Extra configuration parameters to be passed to httr::GET.  Common